+++ src/components/App.scss
@@ -21,7 +21,6 @@ body {
 
   &__theme-dark {
     background-color: #222;
-    color: #fff;
   }
 }
 
@@ -54,12 +53,15 @@ body {
   transition-duration: 0.3s;
 
   &__theme-dark {
-    background-color: #aeaeae;
-    color: #000;
+    background-color: #000;
+    border: 1px solid #008CBA;
+    color: #fff;
   }
 
   &__theme-dark:hover {
-    background-color: #808080;
+    background-color: #000;
+    border: 1px solid #0080B0;
+    color: #0080B0;
   }
 }
 
@@ -75,30 +77,50 @@ body {
   padding: 5px;
   transition-duration: 0.3s;
 
-
   * {
-  margin: 5px;
-  padding: 5px;
+    margin: 5px;
+    padding: 5px;
+    transition-duration: 0.3s;
+  }
+
+  input {
+    border: 1px solid #aaa;
   }
 
   &__description {
-  grid-column: 1/3;
-  resize: none;
-  font-family: 'Roboto', sans-serif;
+    grid-column: 1/3;
+    resize: none;
+    font-family: 'Roboto', sans-serif;
+    border: 1px solid #aaa;
   }
 
   &__submit {
-  grid-column: 1/3;
-  background-color: #008CBA; 
-
+    grid-column: 1/3;
+    background-color: #008CBA; 
   }
 
   &__submit:hover {
-  background-color: #0080B0;
+    background-color: #0080B0;
   }
 
   &__theme-dark {
     background-color: #000;
+  }
+
+  &__theme-dark input  {
+    border: 1px solid #808080;
+    background-color: #000;
+    color: #fff;
+  }
+
+  &__theme-dark textarea  {
+    border: 1px solid #808080;
+    background-color: #000;
+    color: #fff;
+  }
+
+  &__theme-dark *::placeholder  {
+    color: #fff;
   }
 }
 
@@ -110,25 +132,30 @@ body {
   word-wrap: break-word;
   transition-duration: 0.3s;
 
+  p {
+    padding: 10px;
+  }
 
   td, th {
-  padding: 10px;
-  vertical-align: top;
+    vertical-align: top;
+  }
+
+  tr, th {
+    transition-duration: 0.3s;
+  }
+
+  td {
+    padding: 10px;
   }
 
   th {
-  cursor: pointer;
-  text-decoration: underline;
-  border: 1px solid #ddd;
-  transition-duration: 0.3s;
+    cursor: pointer;
+    text-decoration: underline;
+    border: 1px solid #ddd;
   }
 
   th:hover {
-  background-color: #fbfbfb;
-  }
-
-  th:hover {
-  background-color: #fbfbfb;
+    background-color: #fbfbfb;
   }
 
   tr:nth-child(even) {
@@ -144,7 +171,7 @@ body {
   }
 
   th:nth-child(4) {
-    width: 50px;
+    width: 70px;
   }
 
   &__date {
@@ -157,6 +184,18 @@ body {
 
   &__theme-dark {
     background-color: #000;
+  }
+
+  &__theme-dark th {
+    border: 1px solid #808080;
+  }
+
+  &__theme-dark th:hover {
+    background-color: #222;
+  }
+
+  &__theme-dark * {
+    color: #fff;
   }
 
   &__theme-dark tr:nth-child(even) {
